refactor(api/v2): extract Paginated, list params, body envelopes to types.go
Move Paginated[T], ListParams, singleBody[T], singleReadBody[T],
and emptyBody out of labels.go into a shared types.go so future
resources (projects, tasks, ...) reuse them instead of copy-pasting.
Also centralises the pagination math into NewPaginated[T] which
handles the nil-items and total_pages calculation in one place.
